feat: add confirmation prompt for clear cart

+++ src/app/components/handlers.ts
@@ -401,13 +401,16 @@ export async function handlerDefaultAddress(target: HTMLElement): Promise<void> 
 }
 
 export async function clearCart(): Promise<void> {
-  const itemsBtns = document.querySelectorAll('button');
-  itemsBtns.forEach((btn) => {
-    if (btn instanceof HTMLButtonElement) btn.disabled = true;
-  });
-  await deleteCart(JSON.parse(localStorage.cart).id);
-  await createMyCart();
-  setTimeout(() => {
-    getMyCarts();
-  }, 0);
+  const confirm = window.confirm('Are you sure you want to empty the Basket?');
+  if (confirm) {
+    const itemsBtns = document.querySelectorAll('button');
+    itemsBtns.forEach((btn) => {
+      if (btn instanceof HTMLButtonElement) btn.disabled = true;
+    });
+    await deleteCart(JSON.parse(localStorage.cart).id);
+    await createMyCart();
+    setTimeout(() => {
+      getMyCarts();
+    }, 0);
+  }
 }
